Implement on-disk caching.
This caches kernels (with their all their pre-computed discretized versions)
and expensive operations (e.g. convolutions) on disk. This way, if we do a
parameter sweep, we don't need to recalculate when we come back to parameters
that were already used. We save them to disk because this could quickly
eat up RAM.
Summary testing passes. Will check log file when we do parameter
sweeps to see if it is really loading from disk.