Add parameter sweeps and plotting of heat maps
Parameter sweeps are parallelized with ipyparallel. A lot of work has gone
into getting them to interact nicely with the pickling mechanism used to send
functions to the various engines. One hack in srm.py which I am still unable
to explain why it works/is needed.

Data manipulation functions to subsample series.

Many smaller fixes to the caching mechanisms.