Histories and HeatMaps can be reconstructed from raw dumps.
A raw dump is constituted only of NumPy types, and so can be reloaded even
after a data class has been modified. It is also useful for using Theano output
in a non-Theano class, or vice-versa.

Bug fix: a condition in History.compute_up_to made some Theano graphs batch
computable (as they should), but not the equivalent NumPy computations. Not
sure why it was there, but this commit removes it.

Add 'raw' method to History
Add 'from_raw' method to History
Add 'from_raw' method to HeatMaps