Add Spiketrain class, based on sparse arrays. Deprecates Spiketimes.
Spiketimes had tried to solve the problem of sparse spiking data by a list
of deques. This seemed like a good idea because deques allow for O(1) addition
of elements (lists are O(log(n)) and arrays at least as bad). More importantly,
it stored only spiketimes, so convolving with the data was very fast (my very
first dumb implementation with a massive array was painfully slow even for just
a few hundred neurons).
However, the problem with such a sparse representation of data is that you lose
random access to time points, so finding e.g. "the earliest spike after t"
requires a search through each list. Since the list of spiketimes are sorted,
I thought the `np.searchsorted` routine would be fast enough that its O(log(n))
complexity wouldn't show, but that turned out to be wrong. On longer runs, over
90% of the time is spent on the search function, negating any gain from the O(1)
addition of elements.

So I had two options: I could put some optimization hacks the Spiketimes, such
as caching the last returned time and implementing my own search to make use of
this information. Or I could replaced the underlying datatype by a sparse array
(for whatever reason I had not thought of this before), and all the
functionality I needed would be built in. Given that the class is only a few
methods, I chose the latter, as it is pretty much guaranteed to work.