Histories now mostly work with shared variables.
To achieve this, compute_up_to now starts from _original_tidx, rather than
_cur_tidx. For Numpy data, we update now _original_tidx along with _cur_tidx,
so there should be no change. For Theano, we rely on the optimizer not to do
the computation twice.

What this solution does is sidestep the problem of trying to compute updates
when there is nothing to compute (because _cur_tidx is advanced enough). That
means that if we have compiled a function L(tidx), which triggers the
computation of some data A up to tidx, then the following will fail:
    L(100)
    L(100)    # <- fails here
because in the second evaluation there is nothing to compute for A. More
surprising is that
    L(100)
    L(101)
also fails, but
    L(100)
    L(102)
succeeds.

At the moment this is a bit annoying, but at least workable. Perhaps time will
suggest a solution, or otherwise avoid the problem.