Allow Theano histories to be single operation on tarr
Add the 'all' option to `compute_up_to`. This removes any dependence (and
therefore if else conditionals) on the current time index. If the history is
batch computable, it's _data member becomes a single operation applied to the
time array, and does not get added to the list of updates. This makes it
possible to define discretized kernels, working around the current limitation
that Theano graphs only update 1 step at a time.