[GD] Add (untested) raw format to SGD; options for reported likelihood
Raw format should allow saving descent and reloading them later (for
continuation or plotting).
'Reported likelihood' refers to the current likelihood reported during a fit for
monitoring purposes. Added the 'full' option to compute the full likelihood (not
just that of the time step) and made that the default. There is also a period
option, to avoid computing the likelihood at every iteration.