[GD] FitCollection, w/ plotting functions. repr_np support
Add FitCollection, which collects a number of fits together and provides useful
functions, especially for plotting them together.
This required changes to the SGD initialization, to make it more automatic. This
is the same requirement is the same as for repr_np support, so that function is
also added. Support is however only partial: we still need to instantiate a
SurrogateModel separately. Switching to fully name-based indexing (rather than
symbolic variable-based) should remove this limitation.