[hist,bugfix] Fix issue with arithmetics and subclasses of Series
Operations on Series objects used to return a result of type 
`type(self)`. This is not correct in general (e.g. add `Sin` and `Step` 
history functions: the result is neither, but just a generic Series).
Now we always return the base type Series, and if subclasses need 
operations to remain of their type, they can overload operations 
themselves.