[models] Add `eval()` method. Choose better refhist when building graph.
`eval()` applies symbolic updates and removes them from the update 
dictionary. This is useful e.g. when initializing a model.
When building a graph, we now choose the reference index based on the 
most advanced *unlocked* history. This avoids a problem whene the 
reference history is data, and thus its time index is 1000's of steps 
ahead of the other histories.