[models,bugfix] Remove assertion which prevented converted models to dict
Also fixes a typo in __slots__ which prevented JSON export.

I would prefer leaving the assertion there, but as long as models are 
copied when passed as arguments, I'm not sure there's a better solution.
The alternative would be to always do a deep copy, but exactly because 
Pydantic copies so liberally, that's likely a bad idea.