[models:accumulate,fix] Remove cost from shim symbolic updates
By leaving the cost in the updates dictionary, multiple calls to 
accumulate (e.g. to compile multiple cost functions) grew the updates 
dictionary with updates we do not want.
On the second call, graph manipulations are uselessly made on the first 
cost graph as well.
On the third call, the graph manipulations raise an error because of the 
now broken state of the first cost function.