Fix for NumPy 1.19
NumPy has changed how it treats dtype, which when combined with long 
standing bug with dtype.__eq__, prevents using `np.dtype('Tname')` as 
default in BaseModel.

The fix is to set the default to a string, and let `DType` convert it 
during validation. This required changing `Type` to `DType` in the 
kernels, which is more correct anyway.