Model improvements: callable pending functions, concise __repr__
__repr__ previously returned the result of `summarize`. This is useful 
when you want to know about the model, but when it shows up in debugging 
output, it is really bothersome.
Now the informative output is provided by _repr_html_, so __repr__ can 
be the terse expression other functions (like `list`) assume.

PendingUpdateFunction are now callable. This allows the update methods 
defined on a model to be called directly (albeit by providing `self`) – 
not being able to do so was definitely unintuitive behaviour. This 
feature has already been used in a log-likelihood definition .