Partial fix for deserializing nested models.
Still to fix: serialization of connected histories. We need a way to 
reconnect them to the history of the other submodel. Challenge: probably 
this means that the parent model needs to figure in which order to 
deserialize submodels, and add the missing histories in the keywords.

Also connected histories are serialized multiple times – it should 
suffice to store their full dotted name in any model they are connected 
to.