Add `SubmodelParams` placeholder type, for submodel parameters
During class instantiation, this gets replaced with the `Parameters` 
class of the appropriate submodel.
The current solution is a bit of a hack.
A better (but more difficult) solution would be to make Model optionally 
generic, and allow specifying the submodel types as TypeVars.

Also:
- Standardize the structure comments in `Model`
- remove deprecated `BinomialLogpMixin` and `PyMC3Model` classes – they 
date back to v1, and don't apply to v2 since objective functions are no 
longer part of the model.