Add `.info()` method to Model. Better Model docstrings.
The `Model` docstring is appropriate for `Model` itself, because it 
documents how to subclass it (which is the only legitimate use of 
`Model`).
However, this is irrelevant (and confusing) for users of the concrete 
subclasses. So now when `Model` gets subclassed, it detects whether the 
subclass is abstract or concrete, and in the latter case replaces the 
docstring with the output of `.summarize()`. (param names, var names, 
update functions)