[history,models] Add `numeric` property to History and Model
The `numeric` property is meant to facilitate data analysis pipelines
with models integrated with Theano.
Histories are marked as non-symbolic (therefore indexing returns normal
values), but are otherwise identical to their symbolic originals.

This is still WIP: eventually we would like `numeric` to return an
`DataView`, to keep both the symbolic and numeric histories in sync.